refactor: update service provider tests for TenantConfig
